+++ lib/src/identifier.dart
@@ -1,4 +1,0 @@
-const encodeMethodIdentifier = r'$encoder';
-const decodeMethodIdentifier = r'$decoder';
-const serializerIdentifier = r'$dartson';
-const implementationIdentifier = r'_Dartson$impl';