+++ lib/transformers/date_time.dart
@@ -3,6 +3,7 @@ library dartson.transformers.DateTime;
 import 'package:dartson/dartson.dart';
 
 /// A simple DateTime transformer which uses the toString() method.
+@Deprecated('A default implementation is now available using ISO-8601.')
 class DateTimeParser implements TypeTransformer<String, DateTime> {
   const DateTimeParser();
   DateTime decode(String value) =>