+++ test/generator_test.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:test/test.dart';
 import 'package:build_test/build_test.dart';
-import 'package:dartson/src/generator.dart';
+import 'package:dartson/src/generator/generator.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:path/path.dart' as p;